Last large refactoring
