+++ src/gt4py/storage/cartesian/interface.py
@@ -19,6 +19,8 @@ from typing import Any, Optional, Protocol, Sequence, Union, Tuple
 
 import numpy as np
 
+from . import layout, utils as storage_utils
+
 
 try:
     import cupy as cp
@@ -36,9 +38,11 @@ else:
     ArrayLike = Any  # type: ignore[misc]  # assign multiple types in both branches
     DTypeLike = Any  # type: ignore[misc]  # assign multiple types in both branches
 
-from gt4py._core import definitions
-from . import allocators, layout, utils as storage_utils
+from gt4py._core import definitions as core_defs
+from .. import allocators
 
+
+# Protocols
 class GTDimsInterface(Protocol):
     __gt_dims__: Tuple[str, ...]
 
@@ -47,13 +51,13 @@ class GTOriginInterface(Protocol):
     __gt_origin__: Tuple[int, ...]
 
 
-
-
+# Helper functions
 def _error_on_invalid_preset(backend):
     if backend not in layout.REGISTRY:
         raise RuntimeError(f"Storage preset '{backend}' is not registered.")
 
 
+# Public interface
 def empty(
     shape: Sequence[int],
     dtype: DTypeLike = np.float64,
@@ -99,10 +103,11 @@ def empty(
     storage_info = layout.from_name(backend)
     assert storage_info is not None
     if storage_info["device"] == "gpu":
-        device = definitions.Device(definitions.DeviceType.GPU, 0)
+        # TODO: use CUDA or ROCM depending on GT4PY_USE_HIP config setting
+        device = core_defs.Device(core_defs.DeviceType.CUDA, 0)  
         # allocate_f = storage_utils.allocate_gpu
     else:
-        device = definitions.Device(definitions.DeviceType.CPU, 0)
+        device = core_defs.Device(core_defs.DeviceType.CPU, 0)
         # allocate_f = storage_utils.allocate_cpu
 
     aligned_index, shape, dtype, dimensions = storage_utils.normalize_storage_spec(
@@ -115,9 +120,16 @@ def empty(
     layout_map = storage_info["layout_map"](dimensions)
 
     dtype = np.dtype(dtype)
-    dtype = definitions.dtype(dtype)
-    buffer = allocators.allocate(shape, dtype, layout_map, device=device, byte_alignment=alignment, aligned_index=aligned_index)
-    #_, res = allocate_f(shape, layout_map, dtype, alignment * dtype.itemsize, aligned_index)
+    dtype = core_defs.dtype(dtype)
+    buffer = allocators.allocate(
+        shape,
+        dtype,
+        layout_map,
+        device=device,
+        byte_alignment=alignment,
+        aligned_index=aligned_index,
+    )
+    # _, res = allocate_f(shape, layout_map, dtype, alignment * dtype.itemsize, aligned_index)
     return buffer.ndarray
 
 
@@ -347,81 +359,3 @@ def from_array(
         storage[...] = storage_utils.as_numpy(data)
 
     return storage
-
-
-if dace is not None:
-
-    def dace_descriptor(
-        shape: Sequence[int],
-        dtype: DTypeLike = np.float64,
-        *,
-        backend: str,
-        aligned_index: Optional[Sequence[int]] = None,
-        dimensions: Optional[Sequence[str]] = None,
-    ) -> dace.data.Array:
-        """Return a DaCe data descriptor which describes performance-optimal strides and alignment.
-
-        Parameters
-        ----------
-            shape : `Sequence` of `int`
-                The shape of the resulting data descriptor.
-            dtype :  DTypeLike, optional
-                The dtype encoded in the resulting data descriptor.
-
-        Keyword Arguments
-        -----------------
-            backend : `str`
-                The target backend for which the described allocation is optimized.
-            aligned_index: `Sequence` of `int`, optional
-                Indicate the index of the resulting array that most commonly corresponds to the origin of the compute
-                domain. If not passed, it is aligned at the data origin.
-            dimensions: `Sequence` of `str`, optional
-                Indicate the semantic meaning of the dimensions in the data descriptor. Only used for determining
-                optimal strides, the information is not stored.
-
-        Returns
-        -------
-            DaCe data descriptor
-                With strides that encode padding and aligned to provide optimal performance for the given `backend` and
-                `aligned_index`
-
-        Raises
-        -------
-            TypeError
-                If arguments of an unexpected type are specified.
-            ValueError
-                If illegal or inconsistent arguments are specified.
-        """
-        aligned_index, shape, dtype, dimensions = storage_utils.normalize_storage_spec(
-            aligned_index, shape, dtype, dimensions
-        )
-        itemsize = dtype.itemsize
-        storage_info = layout.from_name(backend)
-        assert storage_info is not None
-        layout_map = storage_info["layout_map"](dimensions)
-
-        order_idx = storage_utils.idx_from_order([i for i in layout_map if i is not None])
-        padded_shape = storage_utils.compute_padded_shape(
-            shape, storage_info["alignment"], order_idx
-        )
-
-        strides = storage_utils.strides_from_padded_shape(padded_shape, order_idx, itemsize)
-
-        storage = (
-            dace.StorageType.GPU_Global
-            if storage_info["device"] == "gpu"
-            else dace.StorageType.CPU_Heap
-        )
-        start_offset = int(np.array([aligned_index]) @ np.array([strides]).T) // itemsize
-
-        total_size = int(int(np.array([shape]) @ np.array([strides]).T) // itemsize)
-
-        start_offset = start_offset % storage_info["alignment"]
-        return dace.data.Array(
-            shape=shape,
-            strides=[s // itemsize for s in strides],
-            dtype=dace.typeclass(str(dtype)),
-            storage=storage,
-            total_size=total_size,
-            start_offset=start_offset,
-        )
